Enhancements for multi-system, multi-manager
Allow OEMs to describe the 'core'
system and manager.

In systems that start manifesting components as
managers and systems in their own right,
allow OEM to indicate which is the 'real' one.

Usually, extra Managers and Systems are something that
has been modeled as such, but from a usual user perspective are not
what they would consider BMCs
and systems.

Change-Id: Ib04e146a7fb0e988dac94c148005e07e2b487b12